feat: html sanitise unit test
